# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema de Calificaciones - API",
    "version": "0.2.0",
    "description": "API REST del Sistema de Calificaciones Escolar. Multi-tenencia por escuela. Autenticación mediante cookie de sesión firmada (HMAC-SHA256) o Bearer token (para automatizaciones)."
  },
  "servers": [
    { "url": "/", "description": "Servidor actual" }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Cookie de sesión firmada con HMAC-SHA256."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token para automatizaciones (AGENT_SECRET_TOKEN). Solo para endpoints específicos."
      }
    },
    "schemas": {
...
      "ConfigActividad": {
        "type": "object",
        "properties": {
          "materiaId": { "type": "string" },
          "trimestre": { "type": "integer" },
          "numActividadesCotidianas": { "type": "integer" },
          "numActividadesIntegradoras": { "type": "integer" },
          "tieneExamen": { "type": "boolean" },
          "numExamenes": { "type": "integer" },
          "porcentajeAC": { "type": "number" },
          "porcentajeAI": { "type": "number" },
          "porcentajeExamen": { "type": "number" }
        }
      },
...
    }
  },
  "security": [{ "sessionCookie": [] }],
  "paths": {
...
    "/api/config-actividades": {
      "get": { "summary": "Obtener configuración de actividades", "responses": { "200": { "description": "Configuración" } } },
      "post": {
        "summary": "Guardar configuración de actividades",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigActividad" } } } },
        "responses": { "200": { "description": "Configuración guardada" } }
      }
    },
...
  }
}